Fixed data type issues in models for incoming weather data

+++ app/src/main/java/com/weatherapp/models/Clouds.kt
@@ -3,5 +3,5 @@ package com.weatherapp.models
 import java.io.Serializable
 
 data class Clouds(
-    val all: Int
+    val all: Double
 ) : Serializable
