+++ app/src/main/java/com/weatherapp/models/WeatherResponse.kt
@@ -7,12 +7,12 @@ data class WeatherResponse(
     val weather: List<Weather>,
     val base: String,
     val main: Main,
-    val visibility: Int,
+    val visibility: Double,
     val wind: Wind,
     val clouds: Clouds,
     val dt: Int,
     val sys: Sys,
-    val id: Int,
+    val id: Double,
     val name: String,
-    val cod: Int
+    val cod: Double
 ) : Serializable
